requirements_dev-tf2.7metal0.51.txt
jupyterlab==3.1.14
matplotlib==3.4.3
numpy==1.23.1
pandas==1.3.3
seaborn==0.11.2
scikit-learn==1.0
statsmodels==0.13.2
pytest==6.2.5
testbook==0.4.2
mlflow==1.20.2
parsenvy==3.0.2
black==21.9b0
protobuf==3.20.1
tensorflow-macos==2.7.0
tensorflow-metal==0.5.1
jupyterlab
scikit-learn
Keras
tensorflow-hub
pydot
nibabel
tensorflow_datasets
ipywidgets
mlflow
